Wired up the favorite items button
The previous solution was using a lot of JavaScript, when Rails UJS
can handle posting to the URL for us, we just needed to add
`data-method='post'`. Ultimately we want to change the URL to
/items/1/favorite, but Sean was physically uncomfortable that we
haven't committed and so we are doing that now.

We decided to rename the join table from `favorite_items` to `favorites`
because `current_user.favorite_items` implied that we would get back
the actual items, not the join records. By renaming the model to `favorite`,
we were able to add another `favorite_items` association, which gives back
what you'd expect.

We also restructured the controller code to describe what we were doing,
which is adding an item to the user's favorite items. We explicitly chose
not to know or care about the join model, and are letting Active Record
handle those details for us.